pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mimos"
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.6.0.66",
    "pydantic>=1.10.2",
    "pyzmq>=24.0.1",
    "toml>=0.10.2"
]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/machanirobotics/mimos"
"Bug Tracker" = "https://github.com/machanirobotics/mimos/issues"

[project.optional-dependencies]
dev = [
    "black"
]

[tool.setuptools.dynamic]
version = {attr = "mimos.__version__.__version__"}
